// api/src/acebook/transactional/notification/services/enums.ts
export enum NotifTags {
    TICKET_APPROVAL = "ticket_approval",
    POST_TO_TICKET = "post_to_ticket",
    POST = "post",
    COMMENT = "comment",
    TAGGED_DEPARTMENT = "tagged_department",
    TICKET = "ticket",
    TICKET_APPROVED = "ticket_approved",
    TICKET_DISAPPROVED = "ticket_disapproved",
    STATUS_CHANGED = "status_changed",

    WORK_ON_PROGRESS = "work_on_progress",
    TASK_RECEIVED = "task_received",
    TASK_COMPLETED = "task_completed",
    TASK_FAILED = "task_failed",
    TASK_CLOSED = "task_closed",
    CLOSED = "closed",
    APPROVED = "approved",
    DISAPPROVED = "disapproved",
   
    TASK = "task",
    OPEN = "open",
    TO_RECEIVE = "to_receive",
    RECEIVED = "received",

    WORK = "work",
    SUBMISSION = "submission",


}
// Always add migration if new tag is added